+++ pg/a8s_pg.go
@@ -2,7 +2,9 @@ package pg
 
 import (
 	"fmt"
+	"slices"
 	"strconv"
+	"strings"
 
 	"github.com/anynines/a9s-cli-v2/k8s"
 	"github.com/anynines/a9s-cli-v2/makeup"
@@ -14,8 +16,10 @@ const A8sPGServiceInstanceAPIGroup = "postgresql.anynines.com"
 const A8sPGServiceInstanceAPIGroupLabel = "a8s.a9s/dsi-group=" + A8sPGServiceInstanceAPIGroup
 const A8sPGBackupAPIGroup = "backups.anynines.com"
 const A8sPGBackupKind = "Backup"
+const A8sPGBackupKindPlural = "backups"
 const A8sPGRestoreKind = "Restore"
 const A8sPGServiceInstanceKind = "PostgreSQL"
+const A8sPGServiceInstanceKindPlural = "postgresqls"
 const A8sPGServiceInstanceKindLabel = "a8s.a9s/dsi-kind=" + A8sPGServiceInstanceKind
 const A8sPGLabelPrimary = "a8s.a9s/replication-role=master"
 const A8sPGServiceInstanceNameLabelKey = "a8s.a9s/dsi-name"
@@ -188,4 +192,68 @@ func WaitForPGRestoreToBecomeReady(namespace, name string) {
 	} else {
 		makeup.PrintCheckmark(fmt.Sprintf("The restore with the name %s in namespace %s has been successful.", name, namespace))
 	}
+}
+
+func DoesServiceInstanceExist(namespace, name string) bool {
+	// Ignore the Don't Execute flag
+	unattendedMode := true
+
+	commandElements := make([]string, 0)
+	commandElements = append(commandElements, "get")
+	commandElements = append(commandElements, A8sPGServiceInstanceKindPlural)
+
+	// Namespace
+	commandElements = append(commandElements, "-n")
+	commandElements = append(commandElements, namespace)
+
+	// Output jsonpath
+	commandElements = append(commandElements, "-o=jsonpath={.items[*].metadata.name}")
+
+	cmd, output, err := k8s.Kubectl(unattendedMode, commandElements...)
+
+	if err != nil {
+		makeup.ExitDueToFatalError(err, "Can't kubectl using the command: "+cmd.String())
+	}
+
+	outputString := string(output)
+
+	if outputString == "" {
+		return false
+	}
+
+	instanceNames := strings.Fields(outputString)
+
+	return slices.Contains(instanceNames, name)
+}
+
+func DoesBackupExist(namespace, backupName string) bool {
+	// Ignore the Don't Execute flag
+	unattendedMode := true
+
+	commandElements := make([]string, 0)
+	commandElements = append(commandElements, "get")
+	commandElements = append(commandElements, A8sPGBackupKindPlural)
+
+	// Namespace
+	commandElements = append(commandElements, "-n")
+	commandElements = append(commandElements, namespace)
+
+	// Output jsonpath
+	commandElements = append(commandElements, "-o=jsonpath={.items[*].metadata.name}")
+
+	cmd, output, err := k8s.Kubectl(unattendedMode, commandElements...)
+
+	if err != nil {
+		makeup.ExitDueToFatalError(err, "Can't kubectl using the command: "+cmd.String())
+	}
+
+	outputString := string(output)
+
+	if outputString == "" {
+		return false
+	}
+
+	instanceNames := strings.Fields(outputString)
+
+	return slices.Contains(instanceNames, backupName)
 }